Change rand seed uint32

+++ core/types/block.go
@@ -98,7 +98,7 @@ type Header struct {
 	CommitSignature  [48]byte `json:"signature"        gencodec:"required"`
 	CommitBitmap     []byte   `json:"bitmap"           gencodec:"required"` // Contains which validator signed
 
-	RandSeed       uint64      `json:"randomSeed"`
+	RandSeed       uint32      `json:"randomSeed"`
 	ShardStateHash common.Hash `json:"shardStateRoot"`
 }
 
@@ -452,8 +452,8 @@ func Number(b1, b2 *Block) bool {
 }
 
 // AddRandSeed add random seed into block header
-func (b *Block) AddRandSeed(randSeed int64) {
-	b.header.RandSeed = uint64(randSeed)
+func (b *Block) AddRandSeed(randSeed uint32) {
+	b.header.RandSeed = uint32(randSeed)
 }
 
 // AddShardStateHash add shardStateHash into block header
